refactor: TMDb重複定義の共通化
