+++ supabase/functions/_shared/tmdb/similar_test.ts
@@ -1,25 +1,7 @@
 import { assertEquals } from "jsr:@std/assert";
 import { jsonResponse, withEnv, withMockFetch } from "../test-helpers.ts";
 import { fetchTmdbSimilar } from "./similar.ts";
-import type { TmdbSearchResult } from "./types.ts";
-
-function createMovieResult(
-  overrides: Partial<TmdbSearchResult> & { tmdbId: number; title: string },
-): TmdbSearchResult {
-  return {
-    tmdbId: overrides.tmdbId,
-    tmdbMediaType: "movie",
-    workType: "movie",
-    title: overrides.title,
-    originalTitle: overrides.originalTitle ?? `${overrides.title} Original`,
-    overview: overrides.overview ?? "overview",
-    posterPath: overrides.posterPath ?? null,
-    releaseDate: overrides.releaseDate ?? "2024-01-01",
-    jpWatchPlatforms: overrides.jpWatchPlatforms ?? [],
-    hasJapaneseRelease: overrides.hasJapaneseRelease ?? true,
-    rottenTomatoesScore: overrides.rottenTomatoesScore ?? null,
-  };
-}
+import { createMovieResult } from "./test-fixtures.ts";
 
 Deno.test("fetchTmdbSimilar は source と recommendation の重複を除外する", async () => {
   await withEnv(
